Arcal-Bot/requirements.txt
aiohttp==3.6.2
asn1crypto==1.2.0
async-timeout==3.0.1
attrs==19.3.0
Automat==0.8.0
beautifulsoup4==4.8.1
cffi==1.13.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
discord.py==1.3.4
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
multidict==4.5.2
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pyOpenSSL==19.0.0
service-identity==18.1.0
six==1.12.0
websockets==8.0
websocket_client==0.56.0
yarl==1.3.0
zope.interface==4.6.0
requests==2.22.0
coreapi==2.3.3
